moves legacy character-list code to a legacy file for the sake of example
